Move manage::ManageExtension into trussed-se050-manage crate
Similar to the changes in trussed-staging#19 [0], this patch moves the
extension that was previously defined in the backend crate into a
separate extension crate.

[0] https://github.com/trussed-dev/trussed-staging/pull/19
