+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Use extension crates `trussed-manage` and `trussed-wrap-key-to-file` instead
   of backend crate `trussed-staging`, see [trussed-staging#19][].
+- Move `manage::ManageExtension` into `trussed-se050-manage` crate.
 
 [trussed-staging#19]: https://github.com/trussed-dev/trussed-staging/pull/19
